Fix SwiftLint errors by splitting up NotificationManager
